Move user component to subfolder
